Reorganizing code. Adding documentation
